+++ packages/components/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `Tooltip` can now be nested directly inside a `Popover`
 - `AccordionDisclosure` now supports `SimpleLayoutProps`
 - `Tree` uses text5 as text color